requirements.txt
alembic==1.11.1
altair==5.0.1
arrow==1.2.3
asteroid-filterbanks==0.4.0
attrs==23.1.0
blessed==1.20.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
clip @ git+https://github.com/openai/CLIP.git@a9b1bf5920416aaeaec965c25dd9e8f98c864f16
colorama==0.4.6
croniter==1.3.15
dateutils==0.6.12
decorator==5.1.1
deepdiff==6.3.0
ftfy==6.1.1
gitdb==4.0.10
GitPython==3.1.31
HyperPyYAML==1.2.1
idna==3.4
importlib-metadata==6.7.0
inquirer==3.1.3
Jinja2==3.1.2
joblib==1.3.1
jsonschema==4.17.3
julius==0.2.7
librosa==0.10.0.post2
lightning==2.0.4
lightning-cloud==0.5.37
lightning-utilities==0.8.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
mkl-fft==1.3.6
mkl-random @ file:///C:/Users/dev-admin/mkl/mkl_random_1682977971003/work
mkl-service==2.4.0
numba==0.57.1
numpy @ file:///C:/b/abs_5akk51tu0f/croot/numpy_and_numpy_base_1687466253743/work
optuna==3.2.0
packaging==23.1
pandas==2.0.3
Pillow==9.0.0
pooch==1.6.0
protobuf==4.23.3
pyannote.audio @ https://github.com/pyannote/pyannote-audio/archive/refs/heads/develop.zip#sha256=7e6c55f943512f3999ea5bf551de0faf069ddb01c96584f1d557f94874357066
pyannote.core==5.0.0
pyannote.database==5.0.1
pyannote.metrics==3.2.1
pyannote.pipeline==2.3
pyarrow==12.0.1
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytorch-lightning==2.0.4
pytorch-metric-learning==2.2.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
regex==2023.6.3
requests==2.31.0
rich==13.4.2
ruamel.yaml==0.17.28
scikit-learn==1.3.0
scipy==1.11.1
six==1.16.0
smmap==5.0.0
soundfile==0.12.1
speechbrain==0.5.14
starsessions==1.3.0
streamlit==1.24.0
tenacity==8.2.2
tensorboardX==2.6.1
threadpoolctl==3.1.0
toml==0.10.2
toolz==0.12.0
torch==1.7.1
torch-audiomentations==0.11.0
torch-pitch-shift==1.2.4
torchaudio==2.0.2
torchmetrics==0.11.4
torchvision==0.8.2
tornado==6.3.2
tqdm==4.65.0
typing_extensions @ file:///C:/b/abs_5em9ekwz24/croot/typing_extensions_1686602003259/work
tzdata==2023.3
tzlocal==4.3.1
urllib3==2.0.3
validators==0.20.0
watchdog==3.0.0
wcwidth==0.2.6
zipp==3.15.0
